Introducing a new namespace, Omise\Capabilities (for Capabiliti API).
